requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langchain-community>=0.2.0
streamlit>=1.35.0
faiss-cpu>=1.8.0
requests>=2.31.0
python-dotenv>=1.0.0
sentence-transformers>=2.0.0
